requirements.txt
requests
beautifulsoup4
praw
PyYAML
newspaper3k
python-docx
pdfplumber
wikipedia
datasets
torch
numpy
tqdm
datasets